+++ src/Lab1/vehicles/Truck.java
